schema: move schema_extension's full definition out of schema_fwd.hh
Found by C++23 expert review. Round 2 moved schema_extension's full
class body into schema_fwd.hh (included by 114 files) to spare 5
extension headers from schema.hh - but the class itself needs
<map>/<string>/<seastar/core/future.hh>, so every one of those 114
files now pays for headers schema_fwd.hh never needed before. A
regression relative to the fwd-declare-only header schema_fwd.hh was
before round 2 touched it.

Move the definition to a new schema/schema_extension.hh; schema_fwd.hh
goes back to a plain `class schema_extension;` forward declaration
(matching its pre-round-2 content). schema.hh and the 5 extension
subclass headers now include schema_extension.hh directly instead of
relying on schema_fwd.hh for the complete type.

Also fixes 2 real bugs the same review found in those 5 headers:
db/per_partition_rate_limit_extension.hh used std::map without
including <map> (worked only by accident, transitively via
schema_fwd.hh's now-removed include), and cdc_extension.hh,
tombstone_gc_extension.hh, db/per_partition_rate_limit_extension.hh,
db/tags/extension.hh all throw std::logic_error without <stdexcept>.
